refactor: integrated global list view everywhere
