+++ apps/app/components/core/board-view/index.ts
@@ -1,0 +1,4 @@
+export * from "./all-boards";
+export * from "./board-header";
+export * from "./single-board";
+export * from "./single-issue";
