+++ apps/app/components/core/index.ts
@@ -1,5 +1,8 @@
+export * from "./board-view";
+export * from "./list-view";
 export * from "./bulk-delete-issues-modal";
 export * from "./existing-issues-list-modal";
 export * from "./image-upload-modal";
-export * from "./issues-filter-view";
+export * from "./issues-view-filter";
+export * from "./issues-view";
 export * from "./not-authorized-view";
