+++ apps/app/components/modules/index.ts
@@ -2,7 +2,6 @@ export * from "./select";
 export * from "./sidebar-select";
 export * from "./delete-module-modal";
 export * from "./form";
-export * from "./list-view";
 export * from "./modal";
 export * from "./module-link-modal";
 export * from "./sidebar";
